requirements.txt
absl-py==0.13.0
astunparse==1.6.3
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
firebase==3.0.1
firebase-admin==5.0.2
Flask==1.1.2
Flask-Mail==0.9.1
flatbuffers==1.12
gast==0.3.3
gcloud==0.17.0
google-api-core==1.31.2
google-api-python-client==2.19.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.0
msgpack==1.0.2
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.1
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
